% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_make_all.R
\name{add_make_all}
\alias{add_make_all}
\title{Add MAKE ALL script}
\usage{
add_make_all()
}
\value{
adds 99_make_all.R to R/ if this doesn't
  already exist.
}
\description{
Adds 99_make_clean.R to R/, and defines files
to be autogenerated. Currently defaults running
all prefix-numbered files up to 89 (files in the
90's are assumed to be infrastructure, not analyses).
}
