% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_make_clean.R
\name{add_make_clean}
\alias{add_make_clean}
\title{Add MAKE CLEAN script}
\usage{
add_make_clean(delete_folder_list = c("results", "data"))
}
\arguments{
\item{delete_folder_list}{vector of folders whose
contents are to be deleted (recursively).}
}
\value{
adds 95_make_clean.R to R/ if this doesn't
  already exist.
}
\description{
Adds 95_make_clean.R to R/, and defines directories
to be purged. Currently defaults to purging results/
and data/.
}
