% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moffitt2018.R
\docType{package}
\name{moffitt2018}
\alias{moffitt2018}
\alias{moffitt2018-package}
\title{Infrastructure Tools Package for Moffitt 2018}
\description{
This package collects various scripts and templates
we've found to be useful in organizing projects.
}
