% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/populate_project.R
\name{populate_project}
\alias{populate_project}
\title{Populate R Project with Common Folders}
\usage{
populate_project(folder_list = c("R", "data", "results"))
}
\arguments{
\item{folder_list}{vector of folder names to create.
Defaults to R/, data/, and results/.}
}
\value{
the folders in folder list should be added
  to the working project if these don't already exist.
}
\description{
Before we begin, we want to make sure the folders
we plan to use will be present and consistently
named.
}
\examples{
populate_project()
}
